requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7
SQLAlchemy==2.0.22
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
alembic==1.12.0
importlib-metadata==6.8.0
mako==1.2.4
click==8.1.7
python-dateutil==2.8.2
six==1.16.0
typing-extensions==4.8.0
zipp==3.17.0